+++ docs/guides/changelog.md
@@ -18,6 +18,7 @@ description: |-
 - Add `ise_device_admin_authentication_rule` resource and data source
 - Add `ise_device_admin_authorization_rule` resource and data source
 - Add `ise_device_admin_authorization_exception_rule` resource and data source
+- Add `ise_device_admin_authorization_global_exception_rule` resource and data source
 
 ## 0.1.5